refactor: components 디렉터리 depth 통일
