+++ src/components/CardExpirationInput/CardExpirationInput.styles.ts
@@ -1,5 +1,5 @@
 import styled from "@emotion/styled";
-import BaseInputWrapper from "../@shared/InputWrapper/InputWrapper";
+import BaseInputWrapper from "../InputWrapper/InputWrapper";
 
 const InputWrapper = styled(BaseInputWrapper)`
   color: #04c09e;
